+++ src/thor/isochrone.cc
@@ -257,6 +257,7 @@ void Isochrone::ExpandForward(GraphReader& graphreader,
 
     // Check if the edge is allowed or if a restriction occurs. Get the edge speed.
     uint32_t speed;
+    bool has_traffic = directededge->predicted_speed() || directededge->constrained_flow_speed() > 0;
     if (has_date_time_) {
       // With date time we check time dependent restrictions and access as well as get
       // traffic based speed if it exists
@@ -276,7 +277,7 @@ void Isochrone::ExpandForward(GraphReader& graphreader,
 
     // Compute the cost to the end of this edge
     Cost newcost = pred.cost() + costing_->EdgeCost(directededge, speed) +
-                   costing_->TransitionCost(directededge, nodeinfo, pred);
+                   costing_->TransitionCost(directededge, nodeinfo, pred, has_traffic);
 
     // Check if edge is temporarily labeled and this path has less cost. If
     // less cost the predecessor is updated and the sort cost is decremented
@@ -473,8 +474,10 @@ void Isochrone::ExpandReverse(GraphReader& graphreader,
     }
 
     // Compute the cost to the end of this edge with separate transition cost
+    bool has_traffic =
+        opp_pred_edge->predicted_speed() || opp_pred_edge->constrained_flow_speed() > 0;
     Cost tc = costing_->TransitionCostReverse(directededge->localedgeidx(), nodeinfo, opp_edge,
-                                              opp_pred_edge);
+                                              opp_pred_edge, has_traffic);
     Cost newcost = pred.cost() + costing_->EdgeCost(opp_edge, speed);
     newcost.cost += tc.cost;