+++ src/thor/timedep_reverse.cc
@@ -148,8 +148,10 @@ void TimeDepReverse::ExpandReverse(GraphReader& graphreader,
       continue;
     }
 
+    bool has_traffic =
+        opp_pred_edge->predicted_speed() || opp_pred_edge->constrained_flow_speed() > 0;
     Cost tc = costing_->TransitionCostReverse(directededge->localedgeidx(), nodeinfo, opp_edge,
-                                              opp_pred_edge);
+                                              opp_pred_edge, has_traffic);
     Cost newcost = pred.cost() +
                    costing_->EdgeCost(opp_edge, t2->GetSpeed(directededge, edgeid, seconds_of_week));
     newcost.cost += tc.cost;